requirements-dev.txt
pytest
rdflib
pyyaml